refactor: Member 관련 custom exception 처리 정리
