+++ backend/src/main/java/com/woowacourse/ternoko/common/exception/ExceptionType.java
@@ -19,11 +19,11 @@ public enum ExceptionType {
 
     // Coach 관련 Exception
     COACH_NOT_FOUND(NOT_FOUND, 2001, "번째 코치를 찾을 수 없습니다."),
-    INVALID_INTERVIEW_COACH_ID(null, 2002, "다른 코치의 예약에 접근할 수 없습니다."),
+    INVALID_INTERVIEW_COACH_ID(BAD_REQUEST, 2002, "다른 코치의 예약에 접근할 수 없습니다."),
 
     // Crew 관련 Exception,
     CREW_NOT_FOUND(NOT_FOUND, 3001, "번째 크루를 찾을 수 없습니다."),
-    INVALID_INTERVIEW_CREW_ID(null, 3002, "다른 크루의 예약에 접근할 수 없습니다."),
+    INVALID_INTERVIEW_CREW_ID(BAD_REQUEST, 3002, "다른 크루의 예약에 접근할 수 없습니다."),
 
     // Interview 관련 Exception,
     INTERVIEW_NOT_FOUND(NOT_FOUND, 4001, "번째 면담 예약을 찾을 수 없습니다."),
@@ -43,7 +43,7 @@ public enum ExceptionType {
     INVALID_STATUS_CREATE_COMMENT(BAD_REQUEST, 6001, "해당 면담은 자유로운 한마디를 입력할 수 없습니다."),
     INVALID_STATUS_FIND_COMMENT(BAD_REQUEST, 6002, "해당 면담은 자유로운 한마디를 조회할 수 없습니다."),
     COMMENT_NOT_FOUND(NOT_FOUND, 6003, "번째 코멘트를 찾을 수 없습니다."),
-    INVALID_INTERVIEW_MEMBER_ID(null, 6004, "해당 면담에 해당하지 않는 회원은 접근할 수 없습니다."),
+    INVALID_INTERVIEW_MEMBER_ID(NOT_FOUND, 6004, "해당 면담에 해당하지 않는 회원은 접근할 수 없습니다."),
     INVALID_COMMENT_MEMBER_ID(BAD_REQUEST, 6005, "해당 코멘트에 해당하지 않는 회원은 접근할 수 없습니다."),
     INVALID_COMMENT_INTERVIEW_ID(BAD_REQUEST, 6006, "코멘트에 해당하지 않는 면담입니다."),
