+++ backend/src/main/java/com/woowacourse/ternoko/common/exception/advice/CommonException.java
@@ -1,4 +1,4 @@
-package com.woowacourse.ternoko.common.exception;
+package com.woowacourse.ternoko.common.exception.advice;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
