+++ backend/src/main/java/com/woowacourse/ternoko/common/exception/exception/CommentInvalidException.java
@@ -1,7 +1,7 @@
 package com.woowacourse.ternoko.common.exception.exception;
 
-import com.woowacourse.ternoko.common.exception.CommonException;
 import com.woowacourse.ternoko.common.exception.ExceptionType;
+import com.woowacourse.ternoko.common.exception.advice.CommonException;
 
 public class CommentInvalidException extends CommonException {
     public CommentInvalidException(final ExceptionType exceptionType) {
