+++ backend/src/main/java/com/woowacourse/ternoko/core/domain/interview/InvalidInterviewMemberException.java
@@ -1,7 +1,7 @@
 package com.woowacourse.ternoko.core.domain.interview;
 
-import com.woowacourse.ternoko.common.exception.CommonException;
 import com.woowacourse.ternoko.common.exception.ExceptionType;
+import com.woowacourse.ternoko.common.exception.advice.CommonException;
 import org.springframework.http.HttpStatus;
 
 public class InvalidInterviewMemberException extends CommonException {
