+++ backend/src/test/java/com/woowacourse/ternoko/core/domain/interview/InterviewTest.java
@@ -24,7 +24,6 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.woowacourse.ternoko.common.exception.MemberNotFoundException;
 import com.woowacourse.ternoko.common.exception.exception.InterviewInvalidException;
 import com.woowacourse.ternoko.core.domain.availabledatetime.AvailableDateTime;
 import com.woowacourse.ternoko.core.domain.interview.formitem.FormItem;
@@ -287,7 +286,7 @@ class InterviewTest {
         final Interview interview = getInterview(EDITABLE);
         // when & then
         assertThatThrownBy(() -> interview.findMemberType(CREW2.getId()))
-                .isInstanceOf(MemberNotFoundException.class);
+                .isInstanceOf(InterviewInvalidException.class);
     }
 
     @ParameterizedTest
